Preserve Source L4 incident count type

+++ scripts/source/project-contract-depth-package-layer4.ts
@@ -489,7 +489,7 @@ async function rebuildViews(client: Client): Promise<void> {
       END AS sla_summary,
       COALESCE(scope.scoped_application_count, 0)::bigint AS scoped_application_count,
       COALESCE(scope.critical_application_count, 0)::bigint AS critical_application_count,
-      0::numeric AS cloud_sev1_sev2_incidents,
+      0::int AS cloud_sev1_sev2_incidents,
       NULL::numeric AS avg_cloud_change_failure_rate,
       COALESCE(perf.credit_earned, 0)::numeric AS service_credits_earned,
       COALESCE(perf.credit_claimed, 0)::numeric AS service_credits_claimed,
@@ -595,7 +595,7 @@ async function rebuildViews(client: Client): Promise<void> {
       COALESCE(fin.linked_budget_amount, 0)::numeric AS linked_budget_amount,
       COALESCE(fin.linked_actual_amount, 0)::numeric AS linked_actual_amount,
       COALESCE(fin.linked_budget_lines, 0)::bigint AS linked_budget_lines,
-      COALESCE(op.cloud_sev1_sev2_incidents, 0)::numeric AS cloud_sev1_sev2_incidents,
+      COALESCE(op.cloud_sev1_sev2_incidents, 0)::int AS cloud_sev1_sev2_incidents,
       op.evidence_gap AS operational_evidence_gap,
       0::bigint AS initiative_dependency_count
     FROM source.contract_vendor_360 c
